move dossier components to new folder, refs 347
