+++ src/app/edit-absences/edit-absences-routing.module.ts
@@ -3,7 +3,7 @@ import { Routes, RouterModule } from '@angular/router';
 import { EditAbsencesComponent } from './components/edit-absences/edit-absences.component';
 import { EditAbsencesListComponent } from './components/edit-absences-list/edit-absences-list.component';
 import { EditAbsencesEditComponent } from './components/edit-absences-edit/edit-absences-edit.component';
-import { StudentDossierComponent } from '../shared/components/student-dossier/student-dossier.component';
+import { StudentDossierComponent } from '../shared/components/dossier/student-dossier/student-dossier.component';
 import { ConfirmAbsencesComponent } from '../shared/components/confirm-absences/confirm-absences.component';
 
 const routes: Routes = [
