+++ src/app/presence-control/presence-control-routing.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { dossierRoute } from '../shared/dossier-route';
+import { dossierRoute } from '../shared/components/dossier/dossier-route';
 import { PresenceControlGroupComponent } from './components/presence-control-group/presence-control-group.component';
 import { PresenceControlListComponent } from './components/presence-control-list/presence-control-list.component';
 import { PresenceControlComponent } from './components/presence-control/presence-control.component';
