+++ src/app/shared/components/dossier/dossier-absences/dossier-absences.component.spec.ts
@@ -1,8 +1,8 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { BehaviorSubject, of } from 'rxjs';
 import { buildTestModuleMetadata } from 'src/spec-helpers';
-import { DossierStateService } from '../../services/dossier-state.service';
-import { StudentProfileAbsencesService } from '../../services/student-profile-absences.service';
+import { DossierStateService } from '../../../services/dossier-state.service';
+import { StudentProfileAbsencesService } from '../../../services/student-profile-absences.service';
 import { StudentProfileAbsencesComponent } from '../student-profile-absences/student-profile-absences.component';
 import { StudentProfileEntryHeaderComponent } from '../student-profile-entry-header/student-profile-entry-header.component';
 import { DossierAbsencesComponent } from './dossier-absences.component';
