+++ src/app/shared/components/dossier/dossier-addresses/dossier-addresses.component.scss
@@ -1,4 +1,4 @@
-@import "../../../../bootstrap-variables";
+@import "../../../../../bootstrap-variables";
 
 .profile-entry {
   display: block;
