+++ src/app/shared/components/dossier/dossier-overview/dossier-overview.component.scss
@@ -1,4 +1,4 @@
-@import "../../../../bootstrap-variables";
+@import "../../../../../bootstrap-variables";
 
 .dossier-section {
   display: flex;
