+++ src/app/shared/components/dossier/student-backlink/student-backlink.component.scss
@@ -1,4 +1,4 @@
-@import "../../../../bootstrap-variables";
+@import "../../../../../bootstrap-variables";
 
 :host {
   display: flex;
