+++ src/app/shared/components/dossier/student-dossier/student-dossier.component.ts
@@ -1,10 +1,10 @@
 import { ChangeDetectionStrategy, Component, Inject } from '@angular/core';
 import { map, switchMap } from 'rxjs';
-import { DossierStateService } from '../../services/dossier-state.service';
+import { DossierStateService } from '../../../services/dossier-state.service';
 import {
   StudentProfileBacklink,
   STUDENT_PROFILE_BACKLINK,
-} from '../../tokens/student-profile-backlink';
+} from '../../../tokens/student-profile-backlink';
 
 @Component({
   selector: 'erz-student-dossier',
