+++ src/app/shared/components/dossier/student-profile-absences/student-profile-absences.component.scss
@@ -1,4 +1,4 @@
-@import "../../../../bootstrap-variables";
+@import "../../../../../bootstrap-variables";
 
 .absence-all {
   padding: 0 0 (0.5 * $spacer) $spacer;
