+++ src/app/shared/components/dossier/student-profile-address/student-profile-address.component.ts
@@ -4,8 +4,8 @@ import {
   Input,
   ChangeDetectionStrategy,
 } from '@angular/core';
-import { Student } from '../../models/student.model';
-import { Person } from '../../models/person.model';
+import { Student } from '../../../models/student.model';
+import { Person } from '../../../models/person.model';
 
 @Component({
   selector: 'erz-student-profile-address',
