+++ src/app/shared/components/dossier/student-profile-apprenticeship-company/student-profile-apprenticeship-company.component.scss
@@ -1,0 +1,5 @@
+@import "../../../../../bootstrap-variables";
+
+.subentry {
+  padding-bottom: $spacer;
+}
