+++ src/app/shared/components/dossier/student-profile-entry-header/student-profile-entry-header.component.scss
@@ -1,4 +1,4 @@
-@import "../../../../bootstrap-variables";
+@import "../../../../../bootstrap-variables";
 
 :host {
   display: flex;
