--- src/app/shared/components/student-profile-apprenticeship-company/student-profile-apprenticeship-company.component.scss
@@ -1,5 +1,0 @@
-@import "../../../../bootstrap-variables";
-
-.subentry {
-  padding-bottom: $spacer;
-}
